Fixed compilation warnings.

+++ bootstrap/bootstrap.cpp
@@ -325,7 +325,7 @@ unsigned int parseDebugFlags(int argc, char** argv)
 {
 	unsigned int result = 0;
 
-	for (size_t i = 1; i < argc; ++i)
+	for (size_t i = 1; i < size_t(argc); ++i)
 	{
 		if (strcmp(argv[i], "--debug-parse") == 0)
 			result |= DebugParse;
@@ -346,7 +346,7 @@ unsigned int parseDebugFlags(int argc, char** argv)
 
 unsigned int parseOptimizationLevel(int argc, char** argv)
 {
-	for (size_t i = 1; i < argc; ++i)
+	for (size_t i = 1; i < size_t(argc); ++i)
 		if (argv[i][0] == '-' && argv[i][1] == 'O' && isdigit(argv[i][2]))
 			return atoi(argv[i] + 2);
 
@@ -355,7 +355,7 @@ unsigned int parseOptimizationLevel(int argc, char** argv)
 
 std::string parseTestName(int argc, char** argv)
 {
-	for (size_t i = 1; i < argc; ++i)
+	for (size_t i = 1; i < size_t(argc); ++i)
 		if (argv[i][0] != '-')
 			return argv[i];
